requirements.txt
lmdb
easydict
seaborn
tqdm
numpy==1.23.0
networkx==3.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-lookups-data==1.0.3
pytorch_transformers==1.2.0
transformers==3.1.0
colour==0.1.5
scikit-learn==1.1.2
scipy==1.8.1
sklearn==0.0
Pillow==9.1.1
PyYAML==6.0
ujson==5.4.0
opt-einsum==3.3.0
matplotlib==3.5.3